Refactored files to delegate responsibilities
