+++ algorithms/rr.py
@@ -1,0 +1,4 @@
+from process import Process, ProcessExecutionRecord
+
+def simulate_rr(arriving_processes: [Process], time_quantum: int) -> ([ProcessExecutionRecord], {int : int}):
+    pass
