+++ algorithms/sjf_co.py
@@ -1,0 +1,4 @@
+from process import Process, ProcessExecutionRecord
+
+def simulate_sjf_co(arriving_processes: [Process]) -> ([ProcessExecutionRecord], {int : int}):
+    pass
