+++ output_builder.py
@@ -1,0 +1,4 @@
+from process import ProcessExecutionRecord
+
+def write_output(records: [ProcessExecutionRecord], waiting_times: {int : int}) -> None:
+    pass
