+++ scheduler.py
@@ -1,0 +1,6 @@
+
+from process import Process, ProcessExecutionRecord
+from algorithms.fcfs import simulate_fcfs
+
+def simulate_scheduler(processes: [Process], algorithm: str) -> ([ProcessExecutionRecord], {int, int}):
+    pass
